refactor: rename bottom tab directory
