+++ src/navigation/BottomTabNavigation/bottomTabNavigation.settings.ts
